refactor(queries): move lib adapters under src/queries/lib
