+++ src/queries/lib/group/list-groups.ts
@@ -1,0 +1,5 @@
+import { client } from '@/orpc/client';
+
+export function listGroups(tournamentId: string) {
+  return client.group.list({ tournamentId });
+}
