+++ src/queries/lib/match/list-matches.ts
@@ -1,0 +1,9 @@
+import { client } from '@/orpc/client';
+
+export function listMatchesByGroup(groupId: string) {
+  return client.match.list({ groupId });
+}
+
+export function listMatchesByTournament(tournamentId: string) {
+  return client.match.list({ tournamentId });
+}
